+++ io-classes/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Added `threadLabel` to `MonadThread`
 * Added `MonadLabelledMVar` class.
+* Added `labelMVar` to `Control.Concurrent.Class.MonadMVar.Strict`
+* Added `withTMVar` and `withTMVarAnd` functions.
 
 ### 1.7.0.0